pyproject.toml
[project]
name = "image-stand"
version = "0.1.0"
description = "Image generation API using kie.ai Nano Banana Pro with LangGraph"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "httpx>=0.27.0",
    "pillow>=10.4.0",
    "python-multipart>=0.0.9",
    "python-dotenv>=1.0.1",
    "scikit-image>=0.24.0",
    "numpy>=1.26.0",
    "langgraph>=0.2.0",
    "langchain-core>=0.3.0",
    "pydantic>=2.0.0",
    "streamlit>=1.37.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "ruff>=0.6.0",
]
